Fix copyrights for GNU ELPA
Assign the Copyright to the FSF.

Include various other minor and cosmetic changes:
- Prefer #' to quote function names.
- Fix a few uses of ' in docstrings which can be rendered incorrectly.
- Remove redundant `:group` args.
- Use keyword args for `define-minor-mode` rather than obsolete
  positional arguments.
- Use `lexical-binding`.
- Add a `Copyright` line where missing.

* polymode-classes.el (pm-chunkmode, pm-inner-chunkmode): Use `symbol`
rather than `face` as type, since `face` is not actually defined as a type.

* polymode-compat.el: Use `with-eval-after-load` throughout.

* polymode-core.el: Don't use #' to quote the symbol passed to `advice-add`
because this arg can't be just a function, it has to be a symbol.
(pm-use-cache): Move before first use.

* template/targets/utils.el (polymode-add-deps-to-load-path): Make better
use of `expand-file-name`.

* tests/compat-tests.el: Skip tests if `poly-markdown` is not available,
instead of signaling an error.